Rename weather to nimbus
I had rebased this branch onto master.

I had most of the code up to this point written under the project name
`weather`, but later renamed the project as `nimbus`. In order to claim
that name on https://crates.io, I had to publish that under master,
meaning I had to combine this branch with master, kind of messing up the
history a bit. Here is where I actually decided to rename the project.